Refactor WorldAnalysis
As a result of the last refactoring commit, it was easy to see that the
`#count` method is doing too much.

Breaking the logic up into smaller, more focused methods, makes it much
easier to read the code and follow what’s going on.

This also makes it much easier to test our code. We can now have
multiple tests for each of the different methods that verify that it
works the way that we want it to.

We still get the benefit of not having to repeat the counting logic and
we get a more expressive api for the class.